Integration tests for check lists

+++ src/main/java/com/wjcwleklinski/listservice/service/CheckListQueryService.java
@@ -1,10 +1,13 @@
 package com.wjcwleklinski.listservice.service;
 
+import com.wjcwleklinski.listservice.error.exception.BadRequestException;
+import com.wjcwleklinski.listservice.error.ErrorMessage;
 import com.wjcwleklinski.listservice.model.view.CheckListDetailsView;
 import com.wjcwleklinski.listservice.repository.CheckListRepository;
 import com.wjcwleklinski.listservice.model.entity.CheckList;
 import com.wjcwleklinski.listservice.model.projection.CheckListCollectionProjection;
 import lombok.RequiredArgsConstructor;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -27,6 +30,10 @@ public class CheckListQueryService {
     }
 
     public void deleteCheckList(String listCode) {
+        CheckList checkList = (CheckList) commonService.getByCode(listCode, checkListRepository);
+        if (CollectionUtils.isNotEmpty(checkList.getEntries())) {
+            throw new BadRequestException(ErrorMessage.CHECKLIST_NOT_EMPTY.getMessage());
+        }
         commonService.deleteByCode(listCode, checkListRepository);
     }
 }
